changes for crud operation in solr

+++ ctaas-dis/src/main/java/tv/cadent/ctaas/dis/vo/SolrQueryVo.java
@@ -24,7 +24,7 @@ public class SolrQueryVo {
 	private String pname;
 	private String pdescription;
 	private int pquantity;
-	private int pprice;
+	private float pprice;
 	private String pcategory;
 
 }
